projects/01_fyyur/starter_code/requirements.txt
# babel==2.9.0
# python-dateutil==2.8.2
# flask-moment==1.0.2
# flask-wtf==1.0.1
# flask_sqlalchemy==2.5.1
alembic==1.7.7
Babel==2.9.0
click==8.1.3
colorama==0.4.4
distlib==0.3.4
filelock==3.7.0
Flask==2.1.2
Flask-Migrate==3.1.0
Flask-Moment==1.0.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
itsdangerous==2.1.2
Jinja2==3.0.0
Mako==1.2.0
MarkupSafe==2.1.1
platformdirs==2.5.2
psycopg2==2.9.3
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.36
virtualenv==20.14.1
Werkzeug==2.0.0
WTForms==3.0.1
